Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Is this solving the qp-driver mismatch problem?
> Previously the release 1.0.9 is different from the staging 1.0.9.
> 
> Surprisingly, the released 1.0.9 is the one that actually works in the TS e2e testing.
> If this commit will create a 1.1.0 based on the 1.0.9 staging, it will break the e2e testing.
> Can you verify what is the problem that the 1.0.9 staging is not working?
> 
> 
> Felix

Felix, This was done because someone complained that the 1.1.0 version had gone missing.  I assumed that someone pushed a change, and the 1.1.0 version was never pushed into release.  My assumption was that the current code in the repo was good and just needed the container label change.

However, if what you say is true, then whoever owns this code needs to fix the issue.  (I'm not quite sure how I landed as a committer to this repo, but the code is Greek to me).    Agree with your -1. and I'm going to abandon this change.

Patch-set: 1